tests/main_tests.rs
use std::error::Error;
use tokio::test;
use clap::{Arg, Command};

// Mock function for get_chatgpt_response
async fn mock_get_chatgpt_response(prompt: &str) -> Result<String, Box<dyn Error>> {
    if prompt == "test prompt" {
        Ok("mock response".to_string())
    } else {
        Err("mock error".into())
    }
}

#[test]
async fn test_valid_prompt() {
    let prompt = "test prompt";
    let response = mock_get_chatgpt_response(prompt).await.unwrap();
    assert_eq!(response, "mock response");
}

#[test]
async fn test_invalid_prompt() {
    let prompt = "invalid prompt";
    let result = mock_get_chatgpt_response(prompt).await;
    assert!(result.is_err());
}

#[test]
async fn test_main_with_valid_prompt() {
    let prompt = "test prompt";
    let matches = Command::new("EthosEngine")
        .version("1.0")
        .author("Your Name <your.email@example.com>")
        .about("Generates philosophical banter")
        .arg(
            Arg::new("prompt")
                .short('p')
                .long("prompt")
                .value_name("PROMPT")
                .help("Sets the prompt for the ChatGPT")
                .num_args(1)
                .required(true),
        )
        .get_matches_from(vec!["EthosEngine", "--prompt", prompt]);

    let prompt = matches
        .get_one::<String>("prompt")
        .expect("Prompt is required");

    let response = mock_get_chatgpt_response(prompt).await.unwrap();
    assert_eq!(response, "mock response");
}

#[test]
async fn test_main_with_invalid_prompt() {
    let prompt = "invalid prompt";
    let matches = Command::new("EthosEngine")
        .version("1.0")
        .author("Your Name <your.email@example.com>")
        .about("Generates philosophical banter")
        .arg(
            Arg::new("prompt")
                .short('p')
                .long("prompt")
                .value_name("PROMPT")
                .help("Sets the prompt for the ChatGPT")
                .num_args(1)
                .required(true),
        )
        .get_matches_from(vec!["EthosEngine", "--prompt", prompt]);

    let prompt = matches
        .get_one::<String>("prompt")
        .expect("Prompt is required");

    let result = mock_get_chatgpt_response(prompt).await;
    assert!(result.is_err());
}
